modified:   main.py

+++ main.py
@@ -1,21 +1,26 @@
 import os
+import logging
 from yt_dlp import YoutubeDL
 from telegram import Update
 from telegram.ext import Application, CommandHandler, MessageHandler, filters, ContextTypes
 from telegram.constants import ChatAction
 import re
 
+# Инициализируем логирование
+logging.basicConfig(level=logging.INFO)
+logger = logging.getLogger(__name__)
 
-
-# Получаем токен из переменной окружения (так он безопаснее хранится на сервере)
+# Получаем токен из переменной окружения
 TELEGRAM_TOKEN = os.getenv("TOKEN")
 
-# Директория для временного хранения загруженных аудио файлов (Railway допускает временную директорию /tmp)
+if not TELEGRAM_TOKEN:
+    raise ValueError("Необходимо установить токен Telegram в переменной окружения 'TOKEN'.")
+
+# Директория для временного хранения загруженных аудио файлов
 output_path = '/tmp'
 
-# Функция для предотвразения ошибки нечитаемых символов в назывании ролика
+# Функция для предотвращения ошибки нечитаемых символов в названии ролика
 def sanitize_filename(filename):
-    # Удаляем запрещенные символы
     return re.sub(r'[\\/:"*?<>|]+', '', filename)
 
 # Функция для загрузки аудио с YouTube и конвертации в MP3
@@ -39,7 +44,7 @@ def download_audio(youtube_url):
             audio_file_path = os.path.join(output_path, f"{sanitized_title}.mp3")
             return audio_file_path
     except Exception as e:
-        print(f"Произошла ошибка при загрузке: {e}")
+        logger.error(f"Произошла ошибка при загрузке: {e}")
         return None
 
 # Функция для обработки команды /start
@@ -49,15 +54,25 @@ async def start(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
 # Функция для обработки сообщений с ссылкой
 async def handle_message(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
     youtube_url = update.message.text
+
+    # Проверка на валидность ссылки
+    if not re.match(r'(https?://)?(www\.)?(youtube|youtu|youtube-nocookie)\.(com|be)/.+', youtube_url):
+        await update.message.reply_text("Пожалуйста, отправьте корректную ссылку на YouTube видео.")
+        return
+
     await update.message.reply_text("Загружаю аудио, подождите несколько секунд...")
     await context.bot.send_chat_action(chat_id=update.effective_chat.id, action=ChatAction.UPLOAD_DOCUMENT)
 
     # Загружаем и конвертируем аудио
     audio_file_path = download_audio(youtube_url)
     if audio_file_path:
-        # Отправляем пользователю аудио файл
-        await update.message.reply_audio(audio=open(audio_file_path, 'rb'))
-        os.remove(audio_file_path)  # Удаляем файл после отправки
+        try:
+            await update.message.reply_audio(audio=open(audio_file_path, 'rb'))
+        except Exception as e:
+            logger.error(f"Ошибка при отправке файла: {e}")
+            await update.message.reply_text("Произошла ошибка при отправке аудио.")
+        finally:
+            os.remove(audio_file_path)  # Удаляем файл после отправки
     else:
         await update.message.reply_text("Произошла ошибка при загрузке аудио.")
